Move Nexus endpoint handlers to owned file
Extract the Nexus endpoint operator RPC adapters into a nexus-prefixed frontend file so the Nexus CODEOWNERS rule applies without changing behavior.
